+++ src/codegen.h
@@ -271,7 +271,7 @@ private:
    */
   void emitValIns(OpCode op, unsigned char paramSize, short restSize) {
     assert(op == OpCode::CALL_FUNC || op == OpCode::CALL_METHOD || op == OpCode::CALL_BUILTIN2 ||
-           op == OpCode::ADD_GLOBBING);
+           op == OpCode::ADD_GLOBBING || op == OpCode::NEW_CLOSURE);
     this->curBuilder().append8(static_cast<unsigned char>(op));
     this->curBuilder().append8(paramSize);
 
@@ -304,6 +304,10 @@ private:
   void emitGlobIns(unsigned char paramSize, bool tilde) {
     this->emitValIns(OpCode::ADD_GLOBBING, paramSize, 1);
     this->curBuilder().append8(tilde ? 1 : 0);
+  }
+
+  void emitNewClosureIns(unsigned char capturedSize) {
+    this->emitValIns(OpCode::NEW_CLOSURE, capturedSize, 0);
   }
 
   /**