+++ src/vm.h
@@ -491,8 +491,15 @@ private:
    *                       | offset |   |        |
    */
   static bool prepareFuncCall(DSState &state, unsigned int paramSize) {
-    auto &func = typeAs<FuncObject>(state.stack.peekByOffset(paramSize));
-    return windStackFrame(state, paramSize + 1, paramSize, func.getCode());
+    const DSCode *code;
+    auto *obj = state.stack.peekByOffset(paramSize).get();
+    if (isa<FuncObject>(obj)) {
+      code = &cast<FuncObject>(obj)->getCode();
+    } else {
+      assert(isa<ClosureObject>(obj));
+      code = &cast<ClosureObject>(obj)->getFuncObj().getCode();
+    }
+    return windStackFrame(state, paramSize + 1, paramSize, *code);
   }
 
   /**